Changed communication mechanism between ReactiveContext and the referenced Actor
Now ReactiveContext transforms usual method calls into reactive paradigm by
explicit Promise-passing. A mutable Promise is passed to the actor. When an
actor finishes the computation, it fulfills the promise. We are aware of the
fact that this is usually discouraged, but we really need some way to
transform message-responses into futures, like the AskPattern, but without
timeouts.