Implemented SeedMain, WorkerMain, LocalScavengerApp and DistributedScavengerApp
The crucial part was to get the configuration more or less sane.
Now the user has to modify only the `scavenger.conf` file that should be placed
in every project directory that uses Scavenger.

This file contains the following settings that are relevant for the user:
* IP and port number of the seed node
* Number of threads used by various node types
* Directory where to store the backed up intermediate results

Just to emphasize it: scavenger.conf is not intended to be used as some kind of
"global Scavenger setting". Each project using Scavenger should have it's own
scavenger.conf file.

Now neither seed, nor worker main require any arguments.
These mains can be started with sbt as follows:
$ sbt "runMain scavenger.app.WorkerMain"
$ sbt "runMain scavenger.app.SeedMain"

The client application should implement a separate main in an object that
extends `DistributedScavengerApp`. It can then also be started as follows:
$ sbt "runMain mypackage.MyApp"